Adds various runtime options.

+++ reshade-steam-proton.sh
@@ -26,9 +26,15 @@ LICENSE
     to the part that asks you for the steamid), we can also use it to install ReShade for a specific game, this
     will also make sure ReShade and the shaders are up to date.
     
-    Requires grep, curl, 7z, wget, protontricks.
+    To skip checking for ReShade / shader updates, pass UPDATE_RESHADE=0
     
-    Tested on various DX9 / DX11 games, doesn't load on some games that use a launcher (Final Fantasy X/X-2 HD Remaster for example).
+    To skip installing d3dcompiler_47, pass D3DCOMPILER=0
+    
+    To use a custom dll override (instead of dxgi.dll) pass CUSTOM_OVERRIDE=
+    This will link dxgi.dll to the game using a different name, for example opengl32.dll and set a wine override for it.
+    Example on a opengl game (Like Wolfenstein: The New Order): CUSTOM_OVERRIDE=opengl32 ./reshade-steam-proton.sh
+    
+    Requires grep, curl, 7z, wget, protontricks.
     
     Example usage:
     
@@ -56,6 +62,17 @@ function printErr() {
     exit 1
 }
 MAIN_PATH=${MAIN_PATH:-~/.reshade}
+CUSTOM_OVERRIDE=${CUSTOM_OVERRIDE:-}
+if [[ ! -z $CUSTOM_OVERRIDE ]] && [[ ! $CUSTOM_OVERRIDE =~ ^(opengl32|d3d11)$ ]]; then
+    echo "You have entered '$CUSTOM_OVERRIDE' as the CUSTOM_OVERRIDE, is this correct?"
+    read -p '(y/n): ' ynCheck
+    if ! [[ $ynCheck =~ ^(y|Y|yes|YES)$ ]]; then
+        echo "Exiting."
+        exit 1
+    fi
+fi
+UPDATE_RESHADE=${UPDATE_RESHADE:-1}
+D3DCOMPILER=${D3DCOMPILER:-1}
 RESHADE_PATH="$MAIN_PATH/reshade"
 SEPERATOR="------------------------------------------------------------------------------------------------"
 echo "ReShade installer/updater for Steam and proton on Linux."
@@ -64,10 +81,10 @@ cd "$MAIN_PATH"
 if [[ ! -d reshade-shaders ]]; then
     echo -e "Installing reshade shaders.\n$SEPERATOR"
     git clone --branch master https://github.com/crosire/reshade-shaders || printErr "Unable to clone https://github.com/crosire/reshade-shaders"
-else
+elif [[ $UPDATE_RESHADE -eq 1 ]]; then
     echo -e "Updating reshade shaders.\n$SEPERATOR"
     cd reshade-shaders
-    git pull || exit 1
+    git pull || printErr "Could not update ReShade shaders."
     cd "$MAIN_PATH"
 fi
 echo "$SEPERATOR"
@@ -76,29 +93,30 @@ VERS=0
 if [[ -e VERS ]]; then
     VERS=$(cat VERS)
 fi
-echo -e "Checking for Reshade updates.\n$SEPERATOR"
-RVERS=$(curl -s https://reshade.me | grep -Po "downloads/\S+?\.exe" || exit 1)
-if ! [[ $? -eq 0 ]]; then printErr "Could not fetch ReShade version."; fi
-if [[ $RVERS != $VERS ]]; then
-    echo -e "Updating Reshade."
-    tmpDir=$(mktemp -d || exit 1)
-    cd "$tmpDir"
-    wget -q https://reshade.me/"$RVERS" || printErr "Could not download latest version of ReShade."
-    exeFile="$(find . -name *.exe || exit 1)"
-    if ! [[ $? -eq 0 ]]; then printErr "Download of ReShade exe file failed."; fi
-    7z -y e "$exeFile" 1> /dev/null || printErr "Failed to extract ReShade using 7z."
-    mv *32.dll d3d9.dll
-    mv *64.dll dxgi.dll
-    rm -f "$exeFile"
-    rm -rf "$RESHADE_PATH"/*
-    mv * "$RESHADE_PATH/"
-    cd "$MAIN_PATH"
-    rm -rf VERS
-    echo "$RVERS" > VERS
-    rm -rf "$tmpDir"
+if [[ ! -f reshade/dxgi.dll ]] || [[ $UPDATE_RESHADE -eq 1 ]]; then
+    echo -e "Checking for Reshade updates.\n$SEPERATOR"
+    RVERS=$(curl -s https://reshade.me | grep -Po "downloads/\S+?\.exe" || exit 1)
+    if ! [[ $? -eq 0 ]]; then printErr "Could not fetch ReShade version."; fi
+    if [[ $RVERS != $VERS ]]; then
+        echo -e "Updating Reshade."
+        tmpDir=$(mktemp -d || exit 1)
+        cd "$tmpDir"
+        wget -q https://reshade.me/"$RVERS" || printErr "Could not download latest version of ReShade."
+        exeFile="$(find . -name *.exe || exit 1)"
+        if ! [[ $? -eq 0 ]]; then printErr "Download of ReShade exe file failed."; fi
+        7z -y e "$exeFile" 1> /dev/null || printErr "Failed to extract ReShade using 7z."
+        mv *32.dll d3d9.dll
+        mv *64.dll dxgi.dll
+        rm -f "$exeFile"
+        rm -rf "$RESHADE_PATH"/*
+        mv * "$RESHADE_PATH/"
+        cd "$MAIN_PATH"
+        rm -rf VERS
+        echo "$RVERS" > VERS
+        rm -rf "$tmpDir"
+    fi
 fi
-echo "$SEPERATOR"
-echo -e "Installing d3dcompiler_47 and setting dxgi override. Make sure to run the game at least once for steam to create the appropriate directories.\n$SEPERATOR"
+echo -e "$SEPERATOR\nInstalling d3dcompiler_47 and setting dxgi override. Make sure to run the game at least once for steam to create the appropriate directories.\n$SEPERATOR"
 echo 'Supply the steamid of the game to install d3dcompiler_47 (To find the steamid, run: protontricks -s Name_Of_Game). (Control+c to exit)'
 echo "Enter n to skip this if d3dcompiler_47 is already installed and dxgi override is already set."
 while true; do
@@ -109,13 +127,22 @@ while true; do
 done
 echo "$SEPERATOR"
 if [[ $steamid =~ ^[0-9]*$ ]]; then
-    protontricks $steamid d3dcompiler_47
-    regFile=~/".local/share/Steam/steamapps/compatdata/$steamid/pfx/user.reg"
-    if [[ -e $regFile ]] && [[ $(grep -Po '^"dxgi"="native,builtin"' $regFile) == "" ]]; then
-        sed -i 's/^"\*d3dcompiler_47"="native"/\0\n"dxgi"="native,builtin"\n"d3d9"="native,builtin"/' "$regFile"
+    if [[ $D3DCOMPILER -eq 1 ]]; then
+        protontricks $steamid d3dcompiler_47
     fi
-    if [[ ! -e $regFile ]] || [[ $(grep -Po '^"dxgi"="native,builtin"' $regFile) == "" ]]; then
-        echo -e "Could not modify or find user.reg file: \"$regFile\"\nManually run: protontricks $steamid winecfg\nIn the Libraries tab, Add dxgi.dll and d3d9.dll and make sure they are set to \"native,builtin\"."
+    regFile=~/".local/share/Steam/steamapps/compatdata/$steamid/pfx/user.reg"
+    if [[ -z $CUSTOM_OVERRIDE ]]; then
+        if [[ -e $regFile ]] && [[ $(grep -Po '^"dxgi"="native,builtin"' $regFile) == "" ]]; then
+            echo "Adding dll overrides for d3d9 and dxgi."
+            sed -i 's/^"\*d3dcompiler_47"="native"/\0\n"dxgi"="native,builtin"\n"d3d9"="native,builtin"/' "$regFile"
+        fi
+        if [[ ! -e $regFile ]] || [[ $(grep -Po '^"dxgi"="native,builtin"' $regFile) == "" ]]; then
+            echo -e "Could not modify or find user.reg file: \"$regFile\"\nManually run: protontricks $steamid winecfg\nIn the Libraries tab, Add dxgi.dll and d3d9.dll and make sure they are set to \"native,builtin\"."
+        fi
+    fi
+    if [[ ! -z $CUSTOM_OVERRIDE ]] && [[ -e $regFile ]] && [[ $(grep -Po "^\"$CUSTOM_OVERRIDE\"=\"native,builtin\"" $regFile) == "" ]]; then
+        echo "Adding dll override for $CUSTOM_OVERRIDE."
+        sed -i "s/^\"\*d3dcompiler_47\"=\"native\"/\0\n\"$CUSTOM_OVERRIDE\"=\"native,builtin\"/" "$regFile"
     fi
 fi
 echo -e "Installing reshade to game directory.\n$SEPERATOR"
@@ -134,6 +161,10 @@ while true; do
     fi
 done
 gamePath="$(realpath "$gamePath")"
+if ! -z [[ $CUSTOM_OVERRIDE ]]; then
+    echo "Linking dxgi.dll as $CUSTOM_OVERRIDE.dll."
+    ln -is $(realpath ~/.reshade/reshade/dxgi.dll) "$gamePath/$CUSTOM_OVERRIDE.dll"
+fi
 ln -is $(realpath ~/.reshade/reshade/*) "$gamePath/"
 ln -is $(realpath ~/.reshade/reshade-shaders/Textures) "$gamePath/"
 ln -is $(realpath ~/.reshade/reshade-shaders/Shaders) "$gamePath/"
